Move LoggedInActivity into the com.hedvig.android.app package
