+++ app/app/src/main/kotlin/com/hedvig/app/service/push/senders/PaymentNotificationSender.kt
@@ -7,12 +7,12 @@ import android.net.Uri
 import androidx.core.app.NotificationCompat
 import androidx.core.app.TaskStackBuilder
 import com.google.firebase.messaging.RemoteMessage
+import com.hedvig.android.app.LoggedInActivity
 import com.hedvig.android.core.common.ApplicationScope
 import com.hedvig.android.core.common.android.notification.setupNotificationChannel
 import com.hedvig.android.navigation.core.HedvigDeepLinkContainer
 import com.hedvig.android.notification.core.NotificationSender
 import com.hedvig.android.notification.core.sendHedvigNotification
-import com.hedvig.app.feature.loggedin.ui.LoggedInActivity
 import com.hedvig.app.service.push.getImmutablePendingIntentFlags
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
